+++ cmd/mp4ff-pslister/main.go
@@ -6,10 +6,10 @@ package main
 import (
 	"encoding/hex"
 	"encoding/json"
+	"errors"
 	"flag"
 	"fmt"
 	"io"
-	"log"
 	"os"
 	"strings"
 
@@ -18,89 +18,108 @@ import (
 	"github.com/Eyevinn/mp4ff/mp4"
 )
 
-var usg = `Usage of mp4ff-pslister:
+const (
+	appName = "mp4ff-pslister"
+)
 
-mp4ff-pslister lists parameter sets for AVC/H.264 or HEVC/H.265 from mp4 sample description, bytestream, or hex input.
+var usg = `Usage of %s:
+
+%s lists parameter sets for AVC/H.264 or HEVC/H.265 from mp4 sample description, bytestream, or hex input.
 
 It prints them as hex and in verbose mode it also prints details in JSON format.
 `
 
-var usage = func(msg string) {
-	parts := strings.Split(os.Args[0], "/")
-	name := parts[len(parts)-1]
-	fmt.Fprintf(os.Stderr, "Error: %s\n", msg)
-	fmt.Fprintln(os.Stderr, usg)
-	fmt.Fprintf(os.Stderr, "%s [-v] [-i <mp4File/byte stream file>] [-vps hex] [-sps hex] [-pps hex]  [-codec avc/hevc]\n", name)
-	flag.PrintDefaults()
+type options struct {
+	inFile  string
+	vpsHex  string
+	spsHex  string
+	ppsHex  string
+	codec   string
+	verbose bool
+	version bool
+}
+
+func parseOptions(fs *flag.FlagSet, args []string) (*options, error) {
+	fs.Usage = func() {
+		fmt.Fprintf(os.Stderr, usg, appName, appName)
+		fmt.Fprintf(os.Stderr, "\n%s [options]\n\noptions:\n", appName)
+		fs.PrintDefaults()
+	}
+
+	opts := options{}
+
+	fs.StringVar(&opts.inFile, "i", "", "Input file (mp4 or byte stream) (alternative to sps")
+	fs.StringVar(&opts.codec, "c", "avc", "Codec to parse (avc or hevc)")
+	fs.StringVar(&opts.vpsHex, "vps", "", "VPS in hex format (HEVC only)")
+	fs.StringVar(&opts.spsHex, "sps", "", "SPS in hex format, alternative to infile")
+	fs.StringVar(&opts.ppsHex, "pps", "", "PPS in hex format")
+	fs.BoolVar(&opts.verbose, "v", false, "Verbose output -> details. On for hex input")
+	fs.BoolVar(&opts.version, "version", false, "Get mp4ff version")
+
+	err := fs.Parse(args[1:])
+	return &opts, err
 }
 
 func main() {
-	verbose := flag.Bool("v", false, "Verbose output -> details. On for hex input")
-	inFile := flag.String("i", "", "mp4 or bytestream file")
-	vpsHex := flag.String("vps", "", "VPS in hex format (HEVC only)")
-	spsHex := flag.String("sps", "", "SPS in hex format")
-	ppsHex := flag.String("pps", "", "PPS in hex format")
-	codec := flag.String("c", "avc", "Codec to parse (avc or hevc or auto)")
-	version := flag.Bool("version", false, "Get mp4ff version")
-
-	flag.Parse()
-
-	if *version {
-		fmt.Printf("mp4ff-pslister %s\n", mp4.GetVersion())
-		os.Exit(0)
-	}
-
-	if *inFile == "" && *spsHex == "" {
-		usage("Must specify infile or sps")
+	if err := run(os.Args, os.Stdout); err != nil {
+		fmt.Fprintf(os.Stderr, "error: %v\n", err)
 		os.Exit(1)
-	}
-
-	if *ppsHex != "" && *spsHex == "" {
-		usage("pps needs sps")
-		os.Exit(1)
-	}
-
-	if *vpsHex != "" {
-		*codec = "hevc"
-	}
-
-	if *spsHex != "" {
-		// Don't just print hex again
-		*verbose = true
-	}
-	err := run(*inFile, *vpsHex, *spsHex, *ppsHex, *codec, *version, *verbose)
-	if err != nil {
-		log.Fatalln(err)
 	}
 }
 
-func run(inFile, vpsHex, spsHex, ppsHex, codec string, version, verbose bool) error {
+func run(args []string, stdout io.Writer) error {
+	fs := flag.NewFlagSet(appName, flag.ContinueOnError)
+	o, err := parseOptions(fs, args)
+
+	if err != nil {
+		if errors.Is(err, flag.ErrHelp) {
+			fs.Usage()
+			return nil
+		}
+		return err
+	}
+
+	if o.version {
+		fmt.Fprintf(stdout, "%s %s\n", appName, mp4.GetVersion())
+		return nil
+	}
+
+	if o.inFile == "" && o.spsHex == "" {
+		fs.Usage()
+		return fmt.Errorf("must specify infile or sps in hex format")
+	}
+
+	if o.vpsHex != "" {
+		o.codec = "hevc"
+	}
+
+	if o.spsHex != "" {
+		// Don't just print hex again
+		o.verbose = true
+	}
+
 	var vpsNalus [][]byte
 	var spsNalus [][]byte
 	var ppsNalus [][]byte
 
-	if inFile != "" {
-		ifd, err := os.Open(inFile)
+	if o.inFile != "" {
+		ifd, err := os.Open(o.inFile)
 		if err != nil {
-			return err
+			return fmt.Errorf("could not open file %s: %w", o.inFile, err)
 		}
 		defer ifd.Close()
 		mp4Extensions := []string{".mp4", ".m4v", ".cmfv"}
 		for _, ext := range mp4Extensions {
-			if strings.HasSuffix(inFile, ext) {
-				err := parseMp4File(ifd, codec, verbose)
-				if err != nil {
-					return err
-				}
-				return nil
+			if strings.HasSuffix(o.inFile, ext) {
+				return parseMp4File(stdout, ifd, o.codec, o.verbose)
 			}
 		}
-		// Assume bytestream
+		// Assume bytestream,AnnexB
 		nalus, err := getNalusFromBytestream(ifd)
 		if err != nil {
 			return err
 		}
-		if codec == "avc" {
+		if o.codec == "avc" {
 			for _, nalu := range nalus {
 				switch avc.GetNaluType(nalu[0]) {
 				case avc.NALU_SPS:
@@ -112,13 +131,12 @@ func run(inFile, vpsHex, spsHex, ppsHex, codec string, version, verbose bool) er
 					ppsNalus = append(ppsNalus, nalu)
 				}
 			}
-			printAvcPS(spsNalus, ppsNalus, verbose)
-			return nil
+			return printAvcPS(stdout, spsNalus, ppsNalus, o.verbose)
 		}
 
 		// hevc
 		for _, nalu := range nalus {
-			switch hevc.NaluType(nalu[0]) {
+			switch naluType := hevc.GetNaluType(nalu[0]); naluType {
 			case hevc.NALU_VPS:
 				if len(spsNalus) > 0 {
 					break // VPS coming back again
@@ -128,90 +146,85 @@ func run(inFile, vpsHex, spsHex, ppsHex, codec string, version, verbose bool) er
 				spsNalus = append(spsNalus, nalu)
 			case hevc.NALU_PPS:
 				ppsNalus = append(ppsNalus, nalu)
+			default:
+				// Ignore other NALUs
 			}
 		}
-		printHevcPS(vpsNalus, spsNalus, ppsNalus, verbose)
-		return nil
+		return printHevcPS(stdout, vpsNalus, spsNalus, ppsNalus, o.verbose)
 	}
 	// Now we have hex case left
-	switch codec {
+	switch o.codec {
 	case "avc":
-		spsNalu, err := hex.DecodeString(spsHex)
+		spsNalu, err := hex.DecodeString(o.spsHex)
 		if err != nil {
 			return err
 		}
 		spsNalus = append(spsNalus, spsNalu)
-		if ppsHex != "" {
-			ppsNalu, err := hex.DecodeString(ppsHex)
+		if o.ppsHex != "" {
+			ppsNalu, err := hex.DecodeString(o.ppsHex)
 			if err != nil {
 				return err
 			}
 			ppsNalus = append(ppsNalus, ppsNalu)
 		}
-		printAvcPS(spsNalus, ppsNalus, verbose)
+		return printAvcPS(stdout, spsNalus, ppsNalus, o.verbose)
 	case "hevc":
-		vpsNalu, err := hex.DecodeString(vpsHex)
+		vpsNalu, err := hex.DecodeString(o.vpsHex)
 		if err != nil {
 			return err
 		}
 		vpsNalus = append(vpsNalus, vpsNalu)
-		spsNalu, err := hex.DecodeString(spsHex)
+		spsNalu, err := hex.DecodeString(o.spsHex)
 		if err != nil {
 			return err
 		}
 		if len(spsNalu) > 0 {
 			spsNalus = append(spsNalus, spsNalu)
 		}
-		ppsNalu, err := hex.DecodeString(ppsHex)
+		ppsNalu, err := hex.DecodeString(o.ppsHex)
 		if err != nil {
 			return err
 		}
 		if len(ppsNalu) > 0 {
 			ppsNalus = append(ppsNalus, ppsNalu)
 		}
-		printHevcPS(vpsNalus, spsNalus, ppsNalus, verbose)
+		return printHevcPS(stdout, vpsNalus, spsNalus, ppsNalus, o.verbose)
 	default:
-		return fmt.Errorf("unknown codec %s", codec)
+		return fmt.Errorf("unknown codec %s", o.codec)
 	}
-	return nil
 }
 
 func getNalusFromBytestream(f io.Reader) ([][]byte, error) {
-	byteStream, err := io.ReadAll(f)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	nalus := avc.ExtractNalusFromByteStream(byteStream)
+	fullRaw, err := io.ReadAll(f)
 	if err != nil {
 		return nil, err
 	}
+
+	nalus := avc.ExtractNalusFromByteStream(fullRaw)
 	return nalus, nil
 }
 
-func parseMp4File(r io.Reader, codec string, verbose bool) error {
+func parseMp4File(w io.Writer, r io.Reader, codec string, verbose bool) error {
 	parsedMp4, err := mp4.DecodeFile(r)
 	if err != nil {
 		return fmt.Errorf("DecodeFile: %w", err)
 	}
 
 	var trackID uint32
-	foundPS := false
-	foundCodec := ""
 	if parsedMp4.Moov != nil {
-		trackID, foundCodec, foundPS, err = parseMp4Init(parsedMp4, verbose)
+		foundPS := false
+		foundCodec := ""
+		trackID, foundCodec, foundPS, err = parseMp4Init(w, parsedMp4, verbose)
 		if err != nil {
 			return fmt.Errorf("parseMp4Init: %w", err)
 		}
-		if codec != "" && codec != foundCodec {
-			return fmt.Errorf("codec mismatch: found %s vs specifed %s", foundCodec, codec)
-		}
+		codec = foundCodec
 		if foundPS {
 			return nil
 		}
-		codec = foundCodec
 	}
 	if parsedMp4.IsFragmented() {
-		err = parseMp4Fragment(parsedMp4, trackID, codec, verbose)
+		err = parseMp4Fragment(w, parsedMp4, trackID, codec, verbose)
 		if err != nil {
 			return fmt.Errorf("parseMp4Fragment: %w", err)
 		}
@@ -233,30 +246,28 @@ func parseMp4File(r io.Reader, codec string, verbose bool) error {
 			mdatPayloadStart := mdat.PayloadAbsoluteOffset()
 			offsetInMdatData := uint64(offset) - mdatPayloadStart
 			sampleData := mdat.Data[offsetInMdatData : offsetInMdatData+uint64(size)]
-			fmt.Println(len(sampleData))
 			switch codec {
 			case "avc":
 				spsNalus, ppsNalus := avc.GetParameterSets(sampleData)
 				if len(spsNalus) == 0 {
 					return fmt.Errorf("no AVC SPS found")
 				}
-				printAvcPS(spsNalus, ppsNalus, verbose)
+				return printAvcPS(w, spsNalus, ppsNalus, verbose)
 			case "hevc":
 				vpsNalus, spsNalus, ppsNalus := hevc.GetParameterSets(sampleData)
 				if len(spsNalus) == 0 {
 					return fmt.Errorf("no HEVC SPS found")
 				}
-				printHevcPS(vpsNalus, spsNalus, ppsNalus, verbose)
+				return printHevcPS(w, vpsNalus, spsNalus, ppsNalus, verbose)
 			default:
 				return fmt.Errorf("unknown codec: %s", codec)
 			}
-			break
 		}
 	}
 	return nil
 }
 
-func parseMp4Init(parsedMp4 *mp4.File, verbose bool) (trackID uint32, codec string, foundPS bool, err error) {
+func parseMp4Init(w io.Writer, parsedMp4 *mp4.File, verbose bool) (trackID uint32, codec string, foundPS bool, err error) {
 	for _, trak := range parsedMp4.Moov.Traks {
 		if trak.Mdia.Hdlr.HandlerType == "vide" {
 			stsd := trak.Mdia.Minf.Stbl.Stsd
@@ -269,7 +280,7 @@ func parseMp4Init(parsedMp4 *mp4.File, verbose bool) (trackID uint32, codec stri
 			}
 			trackID = trak.Tkhd.TrackID
 			if verbose {
-				fmt.Printf("Video %s track ID=%d\n", codec, trackID)
+				fmt.Fprintf(w, "Video %s track ID=%d\n", codec, trackID)
 			}
 			switch codec {
 			case "avc":
@@ -278,8 +289,8 @@ func parseMp4Init(parsedMp4 *mp4.File, verbose bool) (trackID uint32, codec stri
 				if len(spsNalus) == 0 {
 					return trackID, codec, false, nil
 				}
-				printAvcPS(spsNalus, ppsNalus, verbose)
-				return trackID, codec, true, nil
+				err := printAvcPS(w, spsNalus, ppsNalus, verbose)
+				return trackID, codec, true, err
 			case "hevc":
 				vpsNalus := stsd.HvcX.HvcC.GetNalusForType(hevc.NALU_VPS)
 				spsNalus := stsd.HvcX.HvcC.GetNalusForType(hevc.NALU_SPS)
@@ -288,18 +299,17 @@ func parseMp4Init(parsedMp4 *mp4.File, verbose bool) (trackID uint32, codec stri
 					return trackID, codec, false, nil
 				}
 				if stsd.HvcX.Type() == "hev1" {
-					fmt.Printf("Warning: found parameter set nalus although there none in sample descriptor hev1\n")
+					fmt.Fprintf(w, "Warning: found parameter set nalus although there none in sample descriptor hev1\n")
 				}
-				printHevcPS(vpsNalus, spsNalus, ppsNalus, verbose)
-				return trackID, codec, true, nil
+				err := printHevcPS(w, vpsNalus, spsNalus, ppsNalus, verbose)
+				return trackID, codec, true, err
 			}
 		}
 	}
 	return 0, codec, false, fmt.Errorf("no parsable video track found")
 }
 
-func parseMp4Fragment(parsedMp4 *mp4.File, trackID uint32, codec string, verbose bool) error {
-	found := false
+func parseMp4Fragment(w io.Writer, parsedMp4 *mp4.File, trackID uint32, codec string, verbose bool) error {
 	if len(parsedMp4.Segments) == 0 || len(parsedMp4.Segments[0].Fragments) == 0 {
 		return fmt.Errorf("no moov or fragment found in mp4 file")
 	}
@@ -317,83 +327,75 @@ func parseMp4Fragment(parsedMp4 *mp4.File, trackID uint32, codec string, verbose
 	fs := samples[0]
 	switch codec {
 	case "avc":
-		found = true
 		spsNalus, ppsNalus := avc.GetParameterSets(fs.Data)
-		printAvcPS(spsNalus, ppsNalus, verbose)
+		return printAvcPS(w, spsNalus, ppsNalus, verbose)
 	case "hevc":
-		found = true
 		vpsNalus, spsNalus, ppsNalus := hevc.GetParameterSets(fs.Data)
-		printHevcPS(vpsNalus, spsNalus, ppsNalus, verbose)
+		return printHevcPS(w, vpsNalus, spsNalus, ppsNalus, verbose)
 	default:
 		return fmt.Errorf("unknown codec: %s", codec)
 	}
-	if !found {
-		return fmt.Errorf("no parameter sets found")
-	}
-	return nil
 }
 
-func printAvcPS(spsNalus, ppsNalus [][]byte, verbose bool) {
+func printAvcPS(w io.Writer, spsNalus, ppsNalus [][]byte, verbose bool) error {
+	if len(spsNalus) == 0 {
+		return fmt.Errorf("no AVC SPS found")
+	}
 	spsMap := make(map[uint32]*avc.SPS)
 	for i, spsNalu := range spsNalus {
 		sps, err := avc.ParseSPSNALUnit(spsNalu, true /*fullVui*/)
 		if err != nil {
-			fmt.Println("Could not parse SPS")
-			return
+			return fmt.Errorf("ParseSPSNALUnit: %w", err)
 		}
-		printPS("SPS", i+1, spsNalu, sps, verbose)
+		printPS(w, "SPS", i+1, spsNalu, sps, verbose)
 		spsMap[sps.ParameterID] = sps
 	}
 	for i, ppsNalu := range ppsNalus {
 		ppsInfo, err := avc.ParsePPSNALUnit(ppsNalu, spsMap)
 		if err != nil {
-			fmt.Println("Could not parse PPS")
-			return
+			return fmt.Errorf("ParsePPSNALUnit: %w", err)
 		}
-		printPS("PPS", i+1, ppsNalu, ppsInfo, verbose)
+		printPS(w, "PPS", i+1, ppsNalu, ppsInfo, verbose)
 	}
 	sps, _ := avc.ParseSPSNALUnit(spsNalus[0], true /*fullVui*/)
-	fmt.Printf("Codecs parameter (assuming avc1) from SPS id %d: %s\n", sps.ParameterID, avc.CodecString("avc1", sps))
+	fmt.Fprintf(w, "Codecs parameter (assuming avc1) from SPS id %d: %s\n", sps.ParameterID, avc.CodecString("avc1", sps))
+	return nil
 }
 
-func printHevcPS(vpsNalus, spsNalus, ppsNalus [][]byte, verbose bool) {
+func printHevcPS(w io.Writer, vpsNalus, spsNalus, ppsNalus [][]byte, verbose bool) error {
 	for i, vps := range vpsNalus {
-		printPS("VPS", i+1, vps, nil, false)
+		printPS(w, "VPS", i+1, vps, nil, false)
 	}
 	spsMap := make(map[uint32]*hevc.SPS)
 	for i, sps := range spsNalus {
 		spsInfo, err := hevc.ParseSPSNALUnit(sps)
 		if err != nil {
-			fmt.Println("Could not parse SPS")
-			return
+			return fmt.Errorf("ParseSPSNALUnit: %w", err)
 		}
-		printPS("SPS", i+1, sps, spsInfo, verbose)
+		printPS(w, "SPS", i+1, sps, spsInfo, verbose)
 		spsMap[uint32(spsInfo.SpsID)] = spsInfo
 	}
 	for i, pps := range ppsNalus {
 		ppsInfo, err := hevc.ParsePPSNALUnit(pps, spsMap)
 		if err != nil {
-			fmt.Println("Could not parse PPS")
-			return
+			return fmt.Errorf("ParsePPSNALUnit: %w", err)
 		}
-		printPS("PPS", i+1, pps, ppsInfo, verbose)
+		printPS(w, "PPS", i+1, pps, ppsInfo, verbose)
 	}
 
 	if len(spsNalus) > 0 {
 		sps, _ := hevc.ParseSPSNALUnit(spsNalus[0])
-		fmt.Printf("Codecs parameter (assuming hvc1) from SPS id %d: %s\n", sps.SpsID, hevc.CodecString("hvc1", sps))
+		fmt.Fprintf(w, "Codecs parameter (assuming hvc1) from SPS id %d: %s\n", sps.SpsID, hevc.CodecString("hvc1", sps))
 	}
+	return nil
 }
 
-func printPS(name string, nr int, ps []byte, psInfo interface{}, verbose bool) {
+func printPS(w io.Writer, name string, nr int, ps []byte, psInfo interface{}, verbose bool) {
 	hexStr := hex.EncodeToString(ps)
 	length := len(hexStr) / 2
-	fmt.Printf("%s %d len %dB: %+v\n", name, nr, length, hexStr)
+	fmt.Fprintf(w, "%s %d len %dB: %+v\n", name, nr, length, hexStr)
 	if verbose && psInfo != nil {
-		jsonPS, err := json.MarshalIndent(psInfo, "", "  ")
-		if err != nil {
-			log.Fatalln(err.Error())
-		}
-		fmt.Printf("%s\n", string(jsonPS))
+		jsonPS, _ := json.MarshalIndent(psInfo, "", "  ")
+		fmt.Fprintf(w, "%s\n", string(jsonPS))
 	}
 }